crypto/x509:check the Key Usage extension
There are two functions in crypto which relate to KU extension checking:
CheckSignature(x509.go 823) and verifyHandshakeSignature(auth.go 22). Unfortunately it's difficult to check the KU inside the func body because KU is not included in the params, not to say the function doesn't know why it was called actually.So we need to check KU before use the certificate PublicKey.

On the other hand, maybe we should add commits to explain the meaning of KU bits as we was starting to check them.